セキュリティの障害を修正

+++ app/controllers/comment_controller.rb
@@ -1,5 +1,10 @@
 class CommentController < ApplicationController
   def create
+    get_user_info
+    unless @user
+      redirect_to root_path
+      return
+    end
     @comment = Comment.new(comment_params)
     if @comment.save then
       redirect_to home_path(@comment.event_id)
@@ -11,7 +16,16 @@ class CommentController < ApplicationController
   end
   
   def destroy
+    get_user_info
+    unless @user
+      redirect_to root_path
+      return
+    end
     comment = Comment.find(params[:id])
+    if comment.uid != @user.uid
+      redirect_to root_path
+      return
+    end
     comment.destroy
     redirect_to home_path(comment.event_id)
   end
